Unused components added to the project
