+++ frontend/front/src/pages/Public/Components/unused-components/README.md
@@ -1,0 +1,5 @@
+# Unused Components
+
+This directory contains components that have been created but are not currently being used in the project.
+
+Please note that these components may be outdated or require updates before integrating them into the project.
