+++ frontend/front/src/pages/Public/Pages/Home/CarrouselHero.js
@@ -3,7 +3,7 @@ import Carousel from "react-material-ui-carousel";
 import { Paper, Box } from "@mui/material";
 import imageHeroFirst from "../../../../assets/images/copywritingImages/Eric-Richards-unedited-3-EDITED.jpg";
 import imageHeroSecond from "../../../../assets/images/copywritingImages/Eric-Richards-unedited-4-EDITED.jpg";
-import HeroPageBgGroundCenter from "./HeroPageBgGroundCenter";
+import HeroPage from "./HeroPage";
 
 function CarrouselHero(props) {
   const [heroHeight, setHeroHeight] = useState("100vh");
@@ -21,7 +21,7 @@ function CarrouselHero(props) {
   const items = [
     {
       component: (
-        <HeroPageBgGroundCenter
+        <HeroPage
           title="Boston Residents Can "
           titleBold="Save Money with Heat Pumps"
           text1="Heat pumps can help you "
@@ -34,7 +34,7 @@ function CarrouselHero(props) {
     },
     {
       component: (
-        <HeroPageBgGroundCenter
+        <HeroPage
           title="BOSTON RESIDENTS CAN "
           titleBold="BENEFIT FROM HEAT PUMPS"
           text1="Heat pumps are effective, continuous, nearly silent and extremely energy efficient."
